% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_aux.PDS.r
\name{block.Sig}
\alias{block.Sig}
\title{Block diagonal spatial covariance matrix}
\usage{
block.Sig(l.pars, coords, reg.ind)
}
\arguments{
\item{l.pars}{?}

\item{coords}{A matrix of x, y coordinates}

\item{reg.ind}{A vector indicating which region each location belongs to (Note: should this be optional?)}
}
\description{
Creates a block diagonal spatial covariance matrix
}

