% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kriging.PDS.r
\name{c.exp}
\alias{c.exp}
\title{Function to compute conditional expectation (i.e. make predictions)}
\usage{
\method{c}{exp}(likfit.obj, X.mon, coords.mon, reg.mon, y, X.pred, coords.pred,
  reg.pred)
}
\arguments{
\item{likfit.obj}{Object output from my.likfit()}

\item{X.mon}{Covariates (i.e., PLS scores) for monitors}

\item{coords.mon}{Monitor coordinates}

\item{reg.mon}{Vector of regions for monitors}

\item{y}{Monitor pollutant concentrations}

\item{X.pred}{Covariates (i.e., PLS scores) for prediction locations}

\item{coords.pred}{Prediction coordinates}

\item{reg.pred}{Vector of regions for prediction sites}
}
\value{
N x 2 matrix, where N is the number of prediction locations;
  1st column contains predictions, 2nd column contains prediction variances
\describe{
  \item{predictions}{Predictions}
  \item{pred.var}{Prediction variances}
}
}
\description{
Function to compute conditional expectation (i.e. make predictions)
}

