% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_criteria.r
\name{fail_low_landuse}
\alias{fail_low_landuse}
\title{Variable Screening: Rare land use types}
\usage{
fail_low_landuse(my.data, my.all.vars)
}
\arguments{
\item{my.data}{See rawdata under PLSK.full}

\item{my.all.vars}{String vector listing candidate variables; often, rownames(my.data)}
}
\value{
String vector listing variables that failed the cleaning check
}
\description{
A function that screens out land use that never makes up more than 10% of the buffer area.  Searches my.all.vars for 
variable names that start with "lu_".
}

