% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_criteria.r
\name{fail_non_zero}
\alias{fail_non_zero}
\title{Variable Screening: All values are zero}
\usage{
fail_non_zero(my.data, my.all.vars)
}
\arguments{
\item{my.data}{See rawdata under PLSK.full}

\item{my.all.vars}{String vector listing candidate variables; often, rownames(my.data)}
}
\value{
String vector listing variables that failed the cleaning check
}
\description{
A function that screens out variables that are 0 for all points
}

