% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_criteria.r
\name{fail_quantile_check}
\alias{fail_quantile_check}
\title{Variable Screening: Sufficient variability}
\usage{
fail_quantile_check(my.data, my.all.vars)
}
\arguments{
\item{my.data}{See rawdata under PLSK.full}

\item{my.all.vars}{String vector listing candidate variables; often, rownames(my.data)}
}
\value{
String vector listing variables that failed the cleaning check
}
\description{
A function that screens out covariates with too little variability (10th percentile == 90th percentile)
}

