% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_criteria.r
\name{fail_skewed_distro}
\alias{fail_skewed_distro}
\title{Variable Screening: Skewness}
\usage{
fail_skewed_distro(my.data, my.all.vars)
}
\arguments{
\item{my.data}{See rawdata under PLSK.full}

\item{my.all.vars}{String vector listing candidate variables; often, rownames(my.data)}
}
\value{
String vector listing variables that failed the cleaning check
}
\description{
A function that screens out covariates that have extremely skewed distributions (maximum normalized value > 10).  Note: -10, too?
}

