% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_aux.PDS.r
\name{inf.beta}
\alias{inf.beta}
\title{Infer Beta from Maximized Covariance Parameters}
\usage{
inf.beta(l.pars, X, coords, reg.ind, data)
}
\arguments{
\item{coords}{A matrix of x, y coordinates}

\item{reg.ind}{A vector indicating which region each location belongs to (Note: should this be optional?)}
}
\description{
Infer Beta from Maximized Covariance Parameters
}

