% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_readdata.r
\name{init.dr}
\alias{init.dr}
\title{Read pollutant request}
\usage{
init.dr(dataloc = "data/", reqnum, poll_list)
}
\arguments{
\item{dataloc}{Directory containing the data}

\item{reqnum}{Request number}

\item{poll_list}{List of pollutants to read}
}
\description{
Function for reading pollution data.
}

