% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kriging.PDS.r
\name{k.pred}
\alias{k.pred}
\title{Auxiliary function for c.exp()}
\usage{
k.pred(reg.dat, dim.X, beta, log.cov.pars)
}
\arguments{
\item{reg.dat}{data}

\item{dim.X}{ncol(X.all)}

\item{beta}{likfit.obj$beta}

\item{log.cov.pars}{log.cov.pars[(3*j-2):(3*j)]}
}
\value{
N x 2 matrix, where N is the number of prediction locations;
  1st column contains predictions, 2nd column contains prediction variances
\describe{
  \item{mp.ind}{stuff}
  \item{y.all}{stuff}
  \item{v.all}{stuff}
}
}
\description{
Auxiliary function for c.exp()
}

