% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_make_object.R
\name{make_modeling_object}
\alias{make_modeling_object}
\title{Make modeling object}
\usage{
make_modeling_object(rawdata, exclude.vars, UK.varnames, pls.comps = 5)
}
\arguments{
\item{rawdata}{See rawdata under PLSK.full}

\item{exclude.vars}{String vector listing variables to exclude from PLS}

\item{UK.varnames}{Sting vector listing variables that will be included for universal kriging and therefore should
not enter into PLS.}

\item{pls.comps}{Integer indicating the number of PLS components.  Default is 5.}
}
\value{
A list named PLS
The PLS object is a list of 19.  In general below, m is the number of monitors.
\describe{
  \item{coefficients}{List of 3-dimensional matrix (number of... monitors plus 32?, y, and 5 components?)}
  \item{scores}{Matrix that is... monitors by 5 components}
  \item{loadings}{Matrix that is number of monitors plus 32 by 5}
  \item{loading.weights}{a similar matrix?}
  \item{Yscores}{Matrix that is number of m x 5}
  \item{Yloadings}{Matrix 1 x 5 that?}
  \item{projection}{366 by 5 matrix}
  \item{Xmeans}{numeric vector}
  \item{Ymeans}{Number indicating?}
  \item{fitted.values}{334  by 5 matrix}
  \item{residuals}{34x1x5 matrix}
  \item{Xvar}{Vector of 5 named numbers}
  \item{Xtotvar}{Number indicating?}
  \item{fit.time}{Named number indicating?}
  \item{ncomp}{Integer.... 5}
  \item{method}{"kernelpls"}
  \item{call}{language plsr(formula = y ~ covars.pls, ncomp = pls.comps, data = model.obj, validation = "none")}
  \item{terms}{Classes 'terms', 'formula' length 3 y ~ covars.pls}
  \item{model}{data frame of y and covars.pls, which is 334 by 366?}
}
}
\description{
Make modeling object.  Executes PLS regression.
}

