% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_blocks.PDS_UKvars.r
\name{make_nat_design_old}
\alias{make_nat_design_old}
\title{Design matrix with national coefficients}
\usage{
make_nat_design_old(my.object, region.vector, my.pls.comps)
}
\arguments{
\item{my.object}{An object containing ?}

\item{region.vector}{A vector indicating which region each location belongs to (Note: should this be optional?)}

\item{my.pls.comps}{Integer indicating the number of pls components}
}
\value{
A matrix with monitor names as rownames and number of columns = numer of regions * (number of UK covars + Number of PLS components)
}
\description{
Make design matrix with national coefficients
}

