% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_blocks.PDS_UKvars.r
\name{make_nat_vector}
\alias{make_nat_vector}
\title{National vector}
\usage{
make_nat_vector(my.object)
}
\arguments{
\item{my.object}{An object containing ?}
}
\value{
A vector of the regions, indicated by 'all'
}
\description{
Make national vector
}

