% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_make_object.R
\name{make_predict_object}
\alias{make_predict_object}
\title{Make prediction object}
\usage{
make_predict_object(rawdata, model.obj, pls.comps)
}
\arguments{
\item{rawdata}{See rawdata under PLSK.full}

\item{pls.comps}{Integer indicating the number of PLS components}
}
\value{
A list
\describe{
  \item{covars.pls}{Stuff}
  \item{PLS$scores}{}
  \item{UK.covars}{}
  \item{monitors}{}
  \item{obs}{}
  \item{coords}{}
  \item{gps}{}
  \item{weco.monitors}{}
  \item{east.monitors}{}
  \item{west.monitors}{}
}
}
\description{
Make prediction object
}

