% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_blocks.PDS_UKvars.r
\name{make_reg_design}
\alias{make_reg_design}
\title{Make Region Design}
\usage{
make_reg_design(my.object, region.vector, my.pls.comps,
  all.regions = c("east", "weco", "west"))
}
\arguments{
\item{my.object}{An object containing ?}

\item{region.vector}{A vector indicating which region each location belongs to (Note: should this be optional?)}

\item{my.pls.comps}{Integer indicating the number of pls components}

\item{all.regions}{A vector of character strings listing all the regions}
}
\value{
A matrix with monitor names as rownames and number of columns = numer of regions * (number of UK covars + Number of PLS components)
}
\description{
Make region design
}

