% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_blocks.PDS_UKvars.r
\name{make_reg_vector}
\alias{make_reg_vector}
\title{Make regional vector}
\usage{
make_reg_vector(my.object)
}
\arguments{
\item{my.object}{An object containing ?}
}
\value{
A matrix with monitor names as rownames and number of columns = numer of regions * (number of UK covars + Number of PLS components)
}
\description{
Make regional vector
}

