% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kriging.PDS.r
\name{my.likfit}
\alias{my.likfit}
\title{Likelihood Function for Regional Variograms}
\usage{
my.likfit(ini.l.pars, X, coords, reg.ind, data, hess = TRUE, trace = 0,
  factr = 1e+07)
}
\arguments{
\item{ini.l.pars}{Length 3xR vector of initial (log) starting values for the covariance
parameters, where "R" is the number of distinct regions; order of starting values
should be e.g. c(tau1,sigma1,rho1,tau2,sigma2,rho2,tau3,sigma3,rho3)}

\item{X}{Matrix of kriging covariates (in our case, PLS scores)}

\item{coords}{Nx2 matrix of coordinates}

\item{reg.ind}{N-vector of distinct kriging regions.  For traditional UK, set reg.ind = rep(1,N)}

\item{data}{Monitor pollutant exposures}

\item{hess}{Logical; whether or not to compute hessian and test whether it is positive-definite.  Defaults to TRUE.
A positive-definite hessian is a good indicator that true log-likelihood max has been found}

\item{trace}{Numeric, 0 by default}

\item{factr}{Numeric, 1e7 by default}
}
\value{
List containing maximized log-covariance parameters; kriging regression covariates;
  maximized log-likelihood; indicator of positive definite hessian, if hess=TRUE
\describe{
  \item{log.cov.pars}{List of vector of named numbers; includes tau, sigma and rho for each region. Maximized log-covariance parameters?}
  \item{beta}{Matrix (instead of vector?) intercept and betas (b1 and b2?) for each region. The kriging regression covariates?}
  \item{max.log.lik}{Numeric; the maximized log-likelihood}
  \item{hess.pd}{Boolean indicating whether the hessian was positive definite}
  \item{hessian}{A [3xR]x[3xR] matrix which is the Hessian.}
}
}
\description{
Likelihood function for regional variograms
}
\details{
Calls inf.beta and optim
}

