% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.cv_UKvars.r
\name{PLSK.cv.alt}
\alias{PLSK.cv.alt}
\title{Function for Running Cross-Validation}
\usage{
PLSK.cv.alt(rawdata, desc.vars, pls.comps, UK.varnames = NULL,
  manual.cv = NULL, factr = 1e+09, verbose = FALSE, my.segments = 10,
  regional = TRUE, kriging = TRUE)
}
\arguments{
\item{rawdata}{A matrix (?) of the rawdata, which must include the following elements: ???  Defaults to NULL.  This will terminate the function call.}

\item{desc.vars}{A list of ?}

\item{pls.comps}{Integer indicating the number of PLS components.}

\item{UK.varnames}{A list of ?}

\item{manual.cv}{Defaults to NULL.  This will?}

\item{factr}{Defaults to 1e10.  What does it do?}

\item{verbose}{Boolean.  Defaults to TRUE.}

\item{my.segments}{Integer; number of segments?}

\item{regional}{Boolean indicating whether model is regional.  Defaults to TRUE.}

\item{kriging}{Boolean indicating whether to krig residuals.  Defaults to TRUE.}
}
\description{
Running cross-validation.
}
\keyword{cross-validation}

