% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.full_UKvars.r
\name{PLSK.full}
\alias{PLSK.full}
\title{Function for Running a Full Model}
\usage{
PLSK.full(rawdata, desc.vars, pls.comps, UK.varnames = NULL, factr = 1e+09,
  verbose = FALSE, regional = TRUE)
}
\arguments{
\item{rawdata}{A data.frame of the rawdata.  See rawdata section for details. Failing to supply rawdata will terminate the function call.}

\item{desc.vars}{A vector of ?}

\item{pls.comps}{An integer between 1 and 5 indicating the number of PLS components.}

\item{UK.varnames}{Defaults to NULL.  This will?}

\item{factr}{Defaults to 1e9.}

\item{verbose}{Defaults to FALSE.  This will?}

\item{regional}{Defaults to TRUE.  This will?}
}
\value{
A list containing a lot of stuff?
\describe{
  \item{ini.l.pars}{Vector of ?.  NOTE: this is hard-coded, can it be removed from the output?}
  \item{region.vec}{List of vector of named strings, indicating which monitors are in which regions}
  \item{b.region.vec}{List of vector of named strings, indicating which monitors are in which regions}
  \item{v.type}{Character indicating variogram type.  Options include "reg" for regional, or "nat" for national.}
  \item{b.type}{Character indicating beta type.  Options include "reg" for regional, or "nat" for national.}
  \item{v.hash}{List of 2 objects: reg:function and nat:function.  These are functions that have something to do with the variogram.}
  \item{b.hash}{List of 2 objects: reg:function and nat:function.  These are functions that have something to do with the betas.}
  \item{model.obj}{List of 13.  The modeling object, generated by make_modeling_object}
  \item{exclude.vars}{List of string vector listing covariates}
  \item{all.vars}{List of string vector listing covariates}
  \item{miss}{List of Boolean vector indicating ?}
  \item{year}{List of integer indicating year of model}
  \item{pollutant}{List of string indicating pollutant}
  \item{invars}{List of string vector, listing all variables passed to modeling process.  NOTE since this is just rownames(rawdata), can it be removed?}
  \item{rawdata}{List of data.frame inluding all original data}
  \item{desc.vars}{List of string vector, listing ?}
  \item{pls.comps}{List of integer indicating number of PLS components}
  \item{UK.varnames}{Stuff}
  \item{factr}{List of ?}
  \item{verbose}{List of boolean; indicates?}
  \item{regional}{List of boolean; if true then regional, if false then single region}
}
}
\description{
Function for running a full model.
}
\details{
This is one of the workhorse functions of this package.
By default, distance variables are log-transformed.  Geographic variables are screened for sufficient variability, etc. and other stuff I'll specify later.
The function is expecting something like the following to be in the rawdata object:
\describe{
  \item{Location-specifying columns}{location_id, native_id, latitude, longitude, monitor_type, lambert_x, lambert_y, state_plane, state, county}
  \item{Pollution concentrations}{pollutant_conc}
}
}
\section{Modeling Object}{

The modeling object (model.obj) is a list of 13
\describe{
  \item{y}{Vector of named numbers (concentrations, I assume)}
  \item{UK.covars}{Stuff}
  \item{covars.pls}{A matrix of something ?}
  \item{PLS}{List of 19}
  \item{obs}{An integer indicating the number of observations (I think)}
  \item{coords}{A data.frame with x and y coordinates (lambert projection)}
  \item{gps}{A data.frame with lat and long coordinates}
  \item{monitors}{String vector listing monitors}
  \item{weco.monitors}{String vector listing monitors in the weco region}
  \item{east.monitors}{String vector listing monitors in the east region}
  \item{west.monitors}{String vector listing monitors in the west region}
  \item{X}{A matrix with rows equal to the number of monitors and columns?}
  \item{parms}{List of 5, generated from my.likfit}
}
}

\section{PLS}{

The PLS object is a list of 19.  In general below, m is the number of monitors.
\describe{
  \item{coefficients}{List of 3-dimensional matrix (number of... monitors plus 32?, y, and 5 components?)}
  \item{scores}{Matrix that is... monitors by 5 components}
  \item{loadings}{Matrix that is number of monitors plus 32 by 5}
  \item{loading.weights}{a similar matrix?}
  \item{Yscores}{Matrix that is number of m x 5}
  \item{Yloadings}{Matrix 1 x 5 that?}
  \item{projection}{366 by 5 matrix}
  \item{Xmeans}{numeric vector}
  \item{Ymeans}{Number indicating?}
  \item{fitted.values}{334  by 5 matrix}
  \item{residuals}{34x1x5 matrix}
  \item{Xvar}{Vector of 5 named numbers}
  \item{Xtotvar}{Number indicating?}
  \item{fit.time}{Named number indicating?}
  \item{ncomp}{Integer.... 5}
  \item{method}{"kernelpls"}
  \item{call}{language plsr(formula = y ~ covars.pls, ncomp = pls.comps, data = model.obj, validation = "none")}
  \item{terms}{Classes 'terms', 'formula' length 3 y ~ covars.pls}
  \item{model}{data frame of y and covars.pls, which is 334 by 366?}
}
}

\section{Parms}{

The parms object is a list of 5, and is generated by my.likfit.  In general below, m is the number of monitors.
\describe{
  \item{log.cov.pars}{List of vector of named numbers; includes tau, sigma and rho for each region (variogram parms?)}
  \item{beta}{Matrix (instead of vector?) intercept and betas (b1 and b2?) for each region}
  \item{max.log.lik}{A list of a number}
  \item{hess.pd}{A list of a boolean indicating whether the hessian was positive definite (?)}
  \item{hessian}{A matrix (9x9) which is the Hessian.}
}
}

\section{To Do List}{

I'm adding this section as a place to track things we might want to do as improvements
\describe{
  \item{}{}
}
}

