% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_polygon.r
\name{poly.int2}
\alias{poly.int2}
\title{Polygon clipping}
\usage{
poly.int2(xpoly, xtest)
}
\arguments{
\item{xpoly}{A matrix of the (x, y) coordinates that are the vertices of the clipping polygon.  Assumes first point = last point.}

\item{xtest}{A matrix of the (x, y) coordinates of the candidate points}
}
\value{
A matrix of the (x, y) coordinates of points within the polygon.
}
\description{
A function for limiting the points to being within a certain polygon?  Returns a matrix of the (x, y) coordinates of points within the polygon.
}

