% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_aux.PDS.r
\name{prof.lik}
\alias{prof.lik}
\title{Profile Likelihood}
\usage{
prof.lik(l.pars, X, coords, reg.ind, data)
}
\arguments{
\item{coords}{A matrix of x, y coordinates}

\item{reg.ind}{A vector indicating which region each location belongs to (Note: should this be optional?)}
}
\description{
Creates a block diagonal spatial covariance matrix
}

