% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_rawdata.r
\name{rawdata_modeling_database}
\alias{rawdata_modeling_database}
\title{Pulling Data from mysql Database}
\usage{
rawdata_modeling_database(pollutant, year)
}
\arguments{
\item{pollutant}{Character string indicating pollutant to pull (options: PM25, O3, NOx, NO2)}

\item{year}{An integer indicating the year of model, e.g. 2013}
}
\description{
Function for pulling data from mysql.  WARNING: has anyone used this in the last 3 years?!
}

