% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_rawdata.r
\name{rawdata_predict_database}
\alias{rawdata_predict_database}
\title{Pulling Data from mysql Database}
\usage{
rawdata_predict_database(cohort, sp)
}
\arguments{
\item{cohort}{Character string indicating locations to pull, e.g. mesa, sisters}

\item{sp}{A charachter string indicating the state plane, e.g. ALzoneE}
}
\description{
Function for pulling data from mysql.  WARNING: has anyone used this in the last 3 years?!
}

