% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmodel_UKvars.R
\name{runcv}
\alias{runcv}
\title{Wrapper Function for Running Cross-Validation}
\usage{
runcv(pls, UK.varnames, rawdata = NULL, desc.vars, factr = 1e+10,
  manual.cv = NULL, ...)
}
\arguments{
\item{UK.varnames}{A list of ?}

\item{rawdata}{A matrix (?) of the rawdata, which must include the following elements: ???  Defaults to NULL.  This will terminate the function call.}

\item{desc.vars}{A list of ?}

\item{factr}{Defaults to 1e10.  What does it do?}

\item{manual.cv}{Defaults to NULL.  This will?}
}
\description{
A wrapper for running cross-validation.
}
\keyword{cross-validation}

