% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmodel_UKvars.R
\name{runfull}
\alias{runfull}
\title{Wrapper Function for Running a Full Model}
\usage{
runfull(pls, UK.varnames, rawdata = NULL, desc.vars, factr = 1e+10,
  randomsample = NULL, ...)
}
\arguments{
\item{pls}{An integer indicating the number of PLS components}

\item{UK.varnames}{A list of varnames to include in universal kriging (e.g. satellite data?)}

\item{rawdata}{A matrix (?) of the rawdata, which must include the following elements: ???  Defaults to NULL.  This will terminate the function call.}

\item{desc.vars}{A list of variable names ?}

\item{factr}{Defaults to 1e10.  What does it do?}

\item{randomsample}{Defaults to NULL.  This will?}
}
\description{
A wrapper for running a full model.  This function calls PLSK.full; see documentation for PLSK.full for more details.
}

