% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_aux.PDS.r
\name{varcov.eff}
\alias{varcov.eff}
\title{Create an exponential covariance matrix}
\usage{
varcov.eff(coords, l.pars)
}
\arguments{
\item{coords}{A matrix of x, y coordinates}

\item{l.pars}{??}
}
\description{
Create an exponential covariace matrix, based on the distance between locations
}

